Kinesis policies should allow "kinesis:DescribeStreamSummary" actions.
Both KinesisCrudPolicy and KinesisStreamReadPolicy should allow "kinesis:DescribeStreamSummary" action as this action is a subset of "kinesis:DescribeStream", which is often more preferable to call. For instance, in my specific use case I want to get the number of stream shards so I can calculate an appropriate backoff between api calls so as not to exceed total write throughput. It is much better to get the stream summary which includes a shard count, than it is to get the full stream description and count the number of shards.